PIM-3551 : move object detacher to storage utils bundle
