+++ src/Akeneo/Bundle/StorageUtilsBundle/Resources/config/doctrine.yml
@@ -3,6 +3,7 @@ parameters:
     doctrine.document_repository.class: Doctrine\ODM\MongoDB\DocumentRepository
     akeneo_storage_utils.doctrine.smart_manager_registry.class: Akeneo\Bundle\StorageUtilsBundle\Doctrine\SmartManagerRegistry
     akeneo_storage_utils.doctrine.table_name_builder.class:     Akeneo\Bundle\StorageUtilsBundle\Doctrine\TableNameBuilder
+    akeneo_storage_utils.doctrine.object_detacher.class:        Akeneo\Bundle\StorageUtilsBundle\Doctrine\ObjectDetacher
     akeneo_storage_utils.event_subscriber.resolve_target_repository.class: Akeneo\Bundle\StorageUtilsBundle\EventSubscriber\ResolveTargetRepositorySubscriber
 
 services:
@@ -21,4 +22,9 @@ services:
         class: %akeneo_storage_utils.doctrine.table_name_builder.class%
         arguments:
             - '@service_container'
+            - '@akeneo_storage_utils.doctrine.smart_manager_registry'
+
+    akeneo_storage_utils.doctrine.object_detacher:
+        class: %akeneo_storage_utils.doctrine.object_detacher.class%
+        arguments:
             - '@akeneo_storage_utils.doctrine.smart_manager_registry'
