+++ src/Pim/Bundle/CatalogBundle/Resources/config/doctrine.yml
@@ -1,6 +1,5 @@
 parameters:
     doctrine.orm.cache.apc.class: Pim\Bundle\CatalogBundle\Doctrine\ArrayApcCache
-    pim_catalog.doctrine.object_detacher.class:     Pim\Bundle\CatalogBundle\Doctrine\Common\Persistence\ObjectDetacher
 
 services:
     # deprecetad: will be removed in 1.4
@@ -14,9 +13,3 @@ services:
     # deprecetad: will be removed in 1.4
     pim_catalog.doctrine.table_name_builder:
         alias: akeneo_storage_utils.doctrine.table_name_builder
-
-    # TODO : new service and class should be move in AkeneoStorageUtils, add specs too
-    pim_catalog.doctrine.object_detacher:
-        class: %pim_catalog.doctrine.object_detacher.class%
-        arguments:
-            - '@pim_catalog.doctrine.smart_manager_registry'
